rename submodule: ml -> encoders
