+++ src/test/test_utils.py
@@ -5,7 +5,7 @@ from cpr_sdk.parser_models import BlockType, ParserOutput, PDFTextBlock
 
 from cli.test.conftest import test_pdf_file_json  # noqa: F401
 from src import config
-from src.ml import SBERTEncoder
+from src.encoders import SBERTEncoder
 from src.utils import (
     encode_parser_output,
     filter_blocks,
